KAFKA-17469: Move persister related classes to persister pkg.
