+++ core/src/test/java/kafka/server/share/SharePartitionTest.java
@@ -37,14 +37,14 @@ import org.apache.kafka.common.record.MemoryRecordsBuilder;
 import org.apache.kafka.common.record.TimestampType;
 import org.apache.kafka.common.utils.MockTime;
 import org.apache.kafka.common.utils.Time;
-import org.apache.kafka.server.share.NoOpShareStatePersister;
-import org.apache.kafka.server.share.PartitionFactory;
-import org.apache.kafka.server.share.Persister;
-import org.apache.kafka.server.share.PersisterStateBatch;
-import org.apache.kafka.server.share.ReadShareGroupStateResult;
-import org.apache.kafka.server.share.TopicData;
-import org.apache.kafka.server.share.WriteShareGroupStateResult;
 import org.apache.kafka.server.share.acknowledge.ShareAcknowledgementBatch;
+import org.apache.kafka.server.share.persister.NoOpShareStatePersister;
+import org.apache.kafka.server.share.persister.PartitionFactory;
+import org.apache.kafka.server.share.persister.Persister;
+import org.apache.kafka.server.share.persister.PersisterStateBatch;
+import org.apache.kafka.server.share.persister.ReadShareGroupStateResult;
+import org.apache.kafka.server.share.persister.TopicData;
+import org.apache.kafka.server.share.persister.WriteShareGroupStateResult;
 import org.apache.kafka.server.storage.log.FetchPartitionData;
 import org.apache.kafka.server.util.FutureUtils;
 import org.apache.kafka.server.util.timer.SystemTimer;
