+++ share/src/main/java/org/apache/kafka/server/share/persister/DeleteShareGroupStateResult.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.server.share;
+package org.apache.kafka.server.share.persister;
 
 import org.apache.kafka.common.message.DeleteShareGroupStateResponseData;
