+++ share/src/main/java/org/apache/kafka/server/share/persister/GroupTopicPartitionData.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.server.share;
+package org.apache.kafka.server.share.persister;
 
 import java.util.List;
 import java.util.Objects;
