+++ share/src/main/java/org/apache/kafka/server/share/persister/PartitionStateData.java
@@ -15,11 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.server.share;
+package org.apache.kafka.server.share.persister;
 
 /**
- * This interface is implemented by classes used to contain the data for a partition in the
- * interface to {@link Persister}.
+ * This interface is implemented by classes used to contain the data for a partition with the state epoch and
+ * start offset in the interface to {@link Persister}.
  */
-public interface PartitionInfoData {
+public interface PartitionStateData extends PartitionInfoData, PartitionIdData {
+    int stateEpoch();
+
+    long startOffset();
 }
