+++ share/src/main/java/org/apache/kafka/server/share/persister/PersisterResult.java
@@ -15,12 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.server.share;
+package org.apache.kafka.server.share.persister;
 
 /**
- * This interface is implemented by classes used to contain the data for a partition with its partition index
- * in the interface to {@link Persister}.
+ * Marker interface for result classes related to the {@link Persister} result classes.
  */
-public interface PartitionIdData extends PartitionInfoData {
-    int partition();
+public interface PersisterResult {
 }
